refactor: editProfile 핸들러 및 데이터 구조 변경 #894
