+++ frontend/src/common/mock/getUserInfo/data.ts
@@ -1,0 +1,8 @@
+import type { UserProfileResponse } from '../../../pages/editProfile/types/userProfile';
+
+export const USER_PROFILE: UserProfileResponse = {
+  name: '홍길동',
+  gender: '남',
+  phoneNumber: '010-1234-5678',
+  image: null,
+};
